improved handling of parition tables

+++ initat/discovery_server/hm_functions.py
@@ -24,6 +24,7 @@ from django.db.models import Q
 from initat.cluster.backbone.models import device, partition, partition_disc, \
     partition_table, partition_fs, lvm_lv, lvm_vg, sys_partition, net_ip, netdevice, \
     netdevice_speed, peer_information
+from initat.cluster.backbone.models.functions import get_related_models
 from initat.snmp.snmp_struct import ResultNode
 from initat.tools import config_tools
 from initat.tools import logging_tools
@@ -181,10 +182,14 @@ class HostMonitoringMixin(object):
                     else:
                         # read previous settings
                         for entry in cur_pt.partition_disc_set.all().values_list(
-                            "partition__mountpoint", "partition__warn_threshold", "partition__crit_threshold"
+                            "partition__mountpoint",
+                            "partition__warn_threshold",
+                            "partition__crit_threshold",
                         ):
                             prev_th_dict[entry[0]] = (entry[1], entry[2])
-                        for entry in cur_pt.lvm_vg_set.all().values_list("lvm_lv__mountpoint", "lvm_lv__warn_threshold", "lvm_lv__crit_threshold"):
+                        for entry in cur_pt.lvm_vg_set.all().values_list(
+                                "lvm_lv__mountpoint", "lvm_lv__warn_threshold", "lvm_lv__crit_threshold"
+                        ):
                             prev_th_dict[entry[0]] = (entry[1], entry[2])
                         if cur_pt.user_created:
                             self.log(
@@ -193,26 +198,14 @@ class HostMonitoringMixin(object):
                             )
                         else:
                             self.log("deleting previous partition_table {}".format(unicode(cur_pt)))
-                            for rel_obj in cur_pt._meta.get_all_related_objects():
-                                if rel_obj.name in [
-                                    "backbone:partition_disc",
-                                    "backbone:lvm_lv",
-                                    "backbone:lvm_vg",
-                                    "backbone:sys_partition",
-                                    # new naming scheme starting from django1.8
-                                    "partition_disc",
-                                    "lvm_lv",
-                                    "lvm_vg",
-                                    "sys_partition",
-                                ]:
-                                    pass
-                                elif rel_obj.name in ["backbone:device", "device"]:
-                                    for ref_obj in rel_obj.model.objects.filter(Q(**{rel_obj.field.name: cur_pt})):
-                                        self.log("cleaning {} of {}".format(rel_obj.field.name, unicode(ref_obj)))
-                                        setattr(ref_obj, rel_obj.field.name, None)
-                                        ref_obj.save()
-                                else:
-                                    raise ValueError("unknown related object {} for partition_info".format(rel_obj.name))
+                            for _dev in get_related_models(cur_pt, detail=True):
+                                for _attr_name in ["act_partition_table", "partition_table"]:
+                                    if getattr(_dev, _attr_name) == cur_pt:
+                                        self.log("clearing attribute {} of {}".format(_attr_name, unicode(_dev)))
+                                        setattr(_dev, _attr_name, None)
+                                        _dev.save(update_fields=[_attr_name])
+                            if get_related_models(cur_pt):
+                                raise SystemError("unable to delete partition {}".format(unicode(cur_pt)))
                             cur_pt.delete()
                         target_dev.act_partition_table = None
                     # fetch partition_fs
